Get commits per tag, total commits and better UI
Structural updates. A lot of improvements and functional enhancements:

* Changed structure of git and models packages in order to simplify its
usage, make more flexible for changes and enhanceable.
* Added tag count and commits per tag.
* During initialization Statistics class counts total number of commits
(actually it takes a while for big repositories, but later it simplifies
time estimation and percentage calculation for other functions).
* Improved printing in user interface, now it prints tables of authors
with their statistical data: added, deleted lines and their total
commits count.
* In the end the results are saved as text (with tables) like
./results/<program_run_date_project_dir>/stats.txt.
* Created and applied TableInterface metaclass for those objects that
have to displayed as tables (for this purpose we need to be guaranteed
that two functions: get_table_row and get_table_headers will be
available in the object. Currently Author and Tag objects use this
interface as they are the only ones that are stringified as a table.

Closes #30
Closes #10
Closes #50