Move files, improve outputs and refactor
Statistics class is intended for not only git but for every type of
statistics. Hence, it is moved outside and now it provides statistics
also about LOC (lines of code).
Closes #57

Abovementioned change also simplified tha approach of argument
parsing and is possible to retrieve any combination of statistics
with shorter, consistent and simpler way.
Closes #59

Runner class is intented for not only git but for general usage. Hence
it is moved in separated folder.
Closes #61

loc/analyzer.py is intended just for line counting (code, comments,
empty). Hence, the name is changed to LineCounter and the analyze
function is renamed to count.
Closes #63

Removed implicit arguments and wildcard imports
Closes #55

Added possibility to combine different extensions in one programming
language and improved loc output
Closes #36