Add multiline comment support for python
Python multiline comments are different from other languages that we
have in known types so far. Its comment start and end looks similar,
also this kind of comment can be representing the string in the file
if it is assigned to the variable. This check we don't have so far and
we will discover multiline comments if they are starting the line (the
whitespace characters are allowed) but if they are assigned to the
variable for example:
MSG = \
 """ THIS IS THE MESSAGE """

the following line with """ enclosings will be counted as a comment.
Solution for this kind of miscountings we currently don't have
implemented but is considered to add in the future.

Closes #101